+++ packages/assets-controller/src/middlewares/ParallelMiddleware.ts
@@ -61,6 +61,17 @@ export function mergeDataResponses(responses: DataResponse[]): DataResponse {
     }
     if (response.updateMode === 'full') {
       merged.updateMode = 'full';
+    } else if (
+      response.updateMode === 'merge' &&
+      merged.updateMode !== 'full'
+    ) {
+      merged.updateMode = 'merge';
+    } else if (
+      response.updateMode === 'update' &&
+      merged.updateMode !== 'full' &&
+      merged.updateMode !== 'merge'
+    ) {
+      merged.updateMode = 'update';
     }
   }
   merged.updateMode ??= 'merge';